# backend/src/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RDGen API - RustDesk Custom Client Builder",
    "version": "1.0.0",
    "description": "API para construção customizada de clientes RustDesk com branding próprio e configurações personalizadas",
    "contact": {
      "name": "API Support",
      "url": "https://rdgen.nextcoreti.com.br"
    }
  },
  "servers": [
    {
      "url": "https://rdgen.nextcoreti.com.br",
      "description": "Servidor de Produção"
    },
    {
      "url": "http://localhost:3001",
      "description": "Servidor de Desenvolvimento"
    }
  ],
...
  "paths": {
...
    "/api/build": {
      "post": {
        "tags": ["Build"],
        "summary": "Iniciar novo build",
        "description": "Cria e inicia um novo build de cliente RustDesk customizado",
        "security": [
          { "ApiKeyAuth": [] },
          { "BearerAuth": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BuildConfig" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Build iniciado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/BuildJob" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimitExceeded" }
        }
      }
    },
    "/api/status/{jobId}": {
      "get": {
        "tags": ["Build"],
        "summary": "Consultar status do build",
        "description": "Retorna o status atual de um build específico",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "ID do job de build"
          }
        ],
        "responses": {
          "200": {
            "description": "Status do build",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/BuildJob" }
                  }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/artifact/{jobId}": {
      "get": {
        "tags": ["Build"],
        "summary": "Download do artefato",
        "description": "Faz download do executável/instalador gerado",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["exe", "msi", "deb", "rpm", "rpm-suse", "appimage", "pkg", "dmg-x64", "dmg-arm64"],
              "default": "exe"
            },
            "description": "Tipo de artefato para download"
          }
        ],
        "responses": {
          "200": {
            "description": "Arquivo binário do artefato",
            "content": {
              "application/octet-stream": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "400": { "description": "Build não completado ainda" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/api/cancel/{jobId}": {
      "post": {
        "tags": ["Build"],
        "summary": "Cancelar build",
        "description": "Cancela um build em andamento",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Build cancelado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/BuildJob" }
                  }
                }
              }
            }
          },
          "400": { "description": "Build já finalizado" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/upload/icon": {
      "post": {
        "tags": ["Upload"],
        "summary": "Upload de ícone",
        "description": "Faz upload do ícone da aplicação (PNG)",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Arquivo PNG (máx 10MB)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ícone processado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "base64": { "type": "string", "description": "Imagem em base64" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/upload/logo": {
      "post": {
        "tags": ["Upload"],
        "summary": "Upload de logo",
        "description": "Faz upload do logo da aplicação (PNG)",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Arquivo PNG (máx 10MB)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logo processado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "base64": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login no painel",
        "description": "Autenticação para acesso ao painel web",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string", "example": "admin" },
                  "password": { "type": "string", "example": "password" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login bem-sucedido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": { "type": "string" },
                        "user": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/me": {
      "get": {
        "tags": ["API Keys"],
        "summary": "Informações da minha API key",
        "description": "Retorna informações sobre a API key utilizada",
        "security": [{ "ApiKeyAuth": [] }],
        "responses": {
          "200": {
            "description": "Informações da API key",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "tenantId": { "type": "string" },
                        "name": { "type": "string" },
                        "rateLimit": { "type": "number" },
                        "buildsToday": { "type": "number" },
                        "expiresAt": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/me/builds": {
      "get": {
        "tags": ["API Keys"],
        "summary": "Meus builds",
        "description": "Lista todos os builds criados com minha API key",
        "security": [{ "ApiKeyAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de builds",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/BuildJob" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/token/generate": {
      "post": {
        "tags": ["Token"],
        "summary": "Gerar token de conexão",
        "description": "Gera um token temporário para conexão automática via RustDesk",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["deviceId", "password"],
                "properties": {
                  "deviceId": {
                    "type": "string",
                    "description": "ID do dispositivo RustDesk",
                    "example": "123456789"
                  },
                  "password": {
                    "type": "string",
                    "description": "Senha temporária do dispositivo",
                    "example": "abc123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token gerado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "token": {
                      "type": "string",
                      "description": "Token de 64 caracteres",
                      "example": "a1b2c3d4..."
                    },
                    "expiresIn": {
                      "type": "number",
                      "description": "Tempo de expiração em segundos",
                      "example": 60
                    },
                    "rustdeskUrl": {
                      "type": "string",
                      "description": "URL para abrir RustDesk com token",
                      "example": "rustdesk://connection/new/123456789?token=a1b2c3d4..."
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Chave de API para autenticação multi-tenant"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token de sessão do painel web"
      }
    },
...
    "responses": {
      "BadRequest": {
        "description": "Requisição inválida",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": false },
                "error": { "type": "string" }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Não autorizado - API key ou token inválido",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": false },
                "error": { "type": "string", "example": "Unauthorized" }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Recurso não encontrado",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": false },
                "error": { "type": "string", "example": "Not found" }
              }
            }
          }
        }
      },
      "RateLimitExceeded": {
        "description": "Limite de requisições excedido",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": false },
                "error": { "type": "string", "example": "Rate limit exceeded" }
              }
            }
          }
        }
      }
    }
  }
}